requirements.txt
Pillow
nose==1.3.3
nose-exclude==0.2.0
requests==2.3.0